+++ packages/builder/src/stores/builder/index.js
@@ -32,6 +32,8 @@ import { rowActions } from "./rowActions"
 import componentTreeNodesStore from "./componentTreeNodes"
 import { appPublished } from "./published"
 
+import { friendlyNamesStore } from "./friendlyNames"
+
 export {
   componentTreeNodesStore,
   layoutStore,
@@ -67,6 +69,7 @@ export {
   snippets,
   rowActions,
   appPublished,
+  friendlyNamesStore,
 }
 
 export const reset = () => {
